# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Freightrek Server API",
    "version": "1.0.0",
    "description": "REST API for Freightrek freight management system with admin authentication and RBAC",
    "contact": {
      "name": "Freightrek Team",
      "email": "support@freightrek.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://freightrekapi.vercel.app",
      "description": "Production (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development (local)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoint"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/route": {
      "post": {
        "summary": "Create a new route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "routeName",
                  "from",
                  "to"
                ],
                "properties": {
                  "routeName": {
                    "type": "string",
                    "example": "Chennai - Bangalore Express"
                  },
                  "from": {
                    "type": "string",
                    "example": "Chennai"
                  },
                  "to": {
                    "type": "string",
                    "example": "Bangalore"
                  },
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "Guindy",
                      "Tambaram"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created successfully"
          },
          "400": {
            "description": "Validation error or route already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      },
      "get": {
        "summary": "Get all routes with pagination and filters",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of routes"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
    "/admin/route/{id}": {
      "get": {
        "summary": "Get route by ID",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route details"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "put": {
        "summary": "Update route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "routeName": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "delete": {
        "summary": "Delete route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route deleted successfully"
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
    "/admin/route/{id}/status": {
      "patch": {
        "summary": "Update route status",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully"
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
    "/admin/route/{id}/branches": {
      "patch": {
        "summary": "Update the branch list of a route (Branch Management)",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "branches"
                ],
                "properties": {
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "Chengalpattu",
                      "Villupuram",
                      "Trichy",
                      "Erode"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branches updated successfully"
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
...
  },
...
}